Support 64-bit variants of cfg.bin; Support JamCrc32 used in 64-bit cfg.bin;

+++ CfgBinEditor/CfgBinEditor/Forms/ConfigurationTreeViewForm.imgui.cs
@@ -255,10 +255,10 @@ namespace CfgBinEditor.Forms
                 case ValueType.String:
                     return $"{value}";
 
-                case ValueType.Int:
+                case ValueType.Long:
                     return isHex ? $"0x{value:X8}" : $"{value}";
 
-                case ValueType.Float:
+                case ValueType.Double:
                     return $"{value}";
 
                 default:
